moving onlyWellDerivs() to MultisegmentWells
it is only used for MultisegmentWells
